Move UpgradeCard to generic directory to be accessible in courseware
