+++ src/generic/upgrade-card/UpgradeCard.scss
